server/app/services/ml_service.py
import torch
import torch.nn as nn
from torchvision import transforms
from PIL import Image
import asyncio
import random
import os
from typing import Dict, Any

class FakeDetectionService:
    """
    Service for detecting AI-generated/fake content using PyTorch models.
    This is a template - you'll need to replace with your actual model.
    """
    
    def __init__(self):
        self.model = None
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model_loaded = False
        
        # Image preprocessing pipeline
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406], 
                               std=[0.229, 0.224, 0.225])
        ])
    
    def load_model(self, model_path: str = None) -> bool:
        """
        Load the PyTorch model for fake detection.
        
        TODO: Replace this with your actual model loading logic.
        For now, this is a placeholder that doesn't load a real model.
        """
        try:
            if model_path and os.path.exists(model_path):
                # TODO: Load your actual trained model
                # self.model = torch.load(model_path, map_location=self.device)
                # self.model.eval()
                pass
            else:
                # For development: Create a dummy model
                # Replace this with your actual model architecture
                self.model = self._create_dummy_model()
            
            self.model_loaded = True
            print(f"Model loaded successfully on device: {self.device}")
            return True
            
        except Exception as e:
            print(f"Error loading model: {e}")
            self.model_loaded = False
            return False
    
    def _create_dummy_model(self):
        """
        Creates a dummy model for development purposes.
        TODO: Replace this with your actual model architecture.
        """
        class DummyModel(nn.Module):
            def __init__(self):
                super().__init__()
                self.classifier = nn.Linear(1000, 2)  # Fake vs Real
            
            def forward(self, x):
                # Return random predictions for demo
                batch_size = x.shape[0] if len(x.shape) > 3 else 1
                return torch.rand(batch_size, 2)
        
        return DummyModel()
    
    def preprocess_image(self, image_path: str) -> torch.Tensor:
        """
        Preprocess image for model input.
        """
        try:
            image = Image.open(image_path).convert('RGB')
            return self.transform(image).unsqueeze(0)  # Add batch dimension
        except Exception as e:
            raise ValueError(f"Error preprocessing image: {e}")
    
    async def predict(self, image_path: str) -> Dict[str, Any]:
        """
        Predict if the image is AI-generated/fake.
        
        Args:
            image_path: Path to the image file
            
        Returns:
            Dictionary containing:
                - is_fake: Boolean indicating if content is fake
                - confidence: Confidence score (0.0 to 1.0)
                - details: Additional analysis details
        """
        
        # Load model if not already loaded
        if not self.model_loaded:
            self.load_model()
        
        try:
            # Preprocess image
            input_tensor = self.preprocess_image(image_path)
            
            # TODO: Replace this section with your actual model inference
            if self.model_loaded and self.model:
                with torch.no_grad():
                    # Move to device
                    input_tensor = input_tensor.to(self.device)
                    
                    # Get prediction
                    outputs = self.model(input_tensor)
                    
                    # Apply softmax to get probabilities
                    probabilities = torch.softmax(outputs, dim=1)
                    
                    # Get fake probability (assuming index 1 is fake)
                    fake_prob = probabilities[0][1].item()
            else:
                # Fallback: Generate random prediction for demo
                fake_prob = random.uniform(0.1, 0.9)
            
            # Determine if fake based on threshold
            is_fake = fake_prob > 0.5
            confidence = fake_prob if is_fake else (1.0 - fake_prob)
            
            # Additional analysis details (placeholder)
            details = {
                "model_version": "v1.0",
                "analysis_method": "CNN-based detection",
                "features_analyzed": ["texture_patterns", "compression_artifacts", "color_distribution"],
                "processing_notes": "Template implementation - replace with actual model"
            }
            
            return {
                "is_fake": is_fake,
                "confidence": confidence,
                "details": str(details)  # Convert to string for database storage
            }
            
        except Exception as e:
            # Return error case with low confidence
            return {
                "is_fake": False,
                "confidence": 0.0,
                "details": f"Error during prediction: {str(e)}"
            }
    
    def get_model_info(self) -> Dict[str, Any]:
        """Get information about the loaded model."""
        return {
            "model_loaded": self.model_loaded,
            "device": str(self.device),
            "model_type": "CNN-based fake detection",
            "version": "v1.0",
            "status": "Template - Replace with actual model"
        }