server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
torch>=2.2.0
torchvision>=0.17.0
numpy==1.24.3
opencv-python==4.8.1.78
aiofiles==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1